# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/story-setting/": {
      "post": {
        "tags": [
          "Standalone Story Setting"
        ],
        "summary": "Create Standalone Story Setting",
        "description": "Create a new story setting independently (before story creation).",
        "operationId": "create_standalone_story_setting_story_setting__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySettingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story-setting/{story_setting_id}": {
      "get": {
        "tags": [
          "Standalone Story Setting"
        ],
        "summary": "Get Standalone Story Setting",
        "description": "Retrieve a story setting by ID.",
        "operationId": "get_standalone_story_setting_story_setting__story_setting_id__get",
        "parameters": [
          {
            "name": "story_setting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Setting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story/": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "List Stories",
        "description": "Get all stories, optionally filtered by user_id.",
        "operationId": "list_stories_story__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryRead"
                  },
                  "title": "Response List Stories Story  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Story"
        ],
        "summary": "Create Story",
        "description": "Create a new story.",
        "operationId": "create_story_story__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story/{story_id}": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Get Story",
        "description": "Get a specific story by ID",
        "operationId": "get_story_story__story_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/story/{story_id}/setting/": {
      "post": {
        "tags": [
          "Story Setting"
        ],
        "summary": "Create Story Setting",
        "description": "Create story setting configuration.",
        "operationId": "create_story_setting_story__story_id__setting__post",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySettingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story/{story_id}/setting/{story_setting_id}": {
      "get": {
        "tags": [
          "Story Setting"
        ],
        "summary": "Get Story Setting",
        "description": "Retrieve a story setting by its ID.",
        "operationId": "get_story_setting_story__story_id__setting__story_setting_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "story_setting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Setting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Story Setting"
        ],
        "summary": "Update Story Setting",
        "description": "Update an existing story setting.",
        "operationId": "update_story_setting_story__story_id__setting__story_setting_id__put",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "story_setting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Setting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySettingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoryCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "avatar_id",
          "story_setting_id"
        ],
        "title": "StoryCreate",
        "description": "Creating a new story."
      },
      "StoryRead": {
        "properties": {
          "story_id": {
            "type": "integer",
            "title": "Story Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "story_id",
          "user_id",
          "avatar_id",
          "story_setting_id",
          "created_at"
        ],
        "title": "StoryRead",
        "description": "Returning story information.\nUsed when displaying story details or resuming a story session."
      },
      "StorySettingCreate": {
        "properties": {
          "setting_prompt": {
            "type": "string",
            "title": "Setting Prompt"
          }
        },
        "type": "object",
        "required": [
          "setting_prompt"
        ],
        "title": "StorySettingCreate",
        "description": "Creating a story setting/prompt."
      },
      "StorySettingRead": {
        "properties": {
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          },
          "setting_prompt": {
            "type": "string",
            "title": "Setting Prompt"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "story_setting_id",
          "setting_prompt",
          "created_at"
        ],
        "title": "StorySettingRead",
        "description": "Returning story setting information.\nUsed when displaying story configuration/setting details."
      },
      "StorySettingUpdate": {
        "properties": {
          "setting_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setting Prompt"
          }
        },
        "type": "object",
        "title": "StorySettingUpdate",
        "description": "Updating story setting information.\nUsed when user wants to modify the story premise/setting.\n\nOptional"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}